correccion de redireccionamiento en nuevaventa

+++ acciones/ventas_acciones.php
@@ -4,16 +4,8 @@
     $venCon = new ventasControl();
     $data = isset($_POST)? $_POST:'';
 
-    echo "<pre>";
-    var_dump($data);
-    echo "</pre>";
-    if ( array_key_exists("fk_tipo_comprobante", $data)){
-        $data['fk_tipo_comprobante']= 1;
-    }else{
-        $data['fk_tipo_comprobante']= 2;
-    }
 
-  /*  if(isset($_POST)&& array_key_exists('colocar id de venta', $_POST)){
+    if(isset($_POST)&& array_key_exists('colocar id de venta', $_POST)){
         $data = $_POST;
         echo "<br>Entro en el primer if";
         if ($data['id_venta']!= null && $data['id_venta'] != '' && !isset($data['del'])){
@@ -42,11 +34,9 @@
             }
         } 
         
-    }*/
+    }
 
-    echo "<pre>";
-    var_dump($data);
-    echo "</pre>";
-
-    echo "<br/><br/><br/><a href='../Vistas/NuevaVenta.php'>Ir a Nueva Venta</a>"
+    //echo "<br/><br/><br/><a href='../Vistas/NuevaVenta.php'>Ir a Nueva Venta</a>"
+    header("Location: ../Vistas/NuevaVenta.php");
+    die();
 ?>
